Allow overriding VSCode telemetry config
