+++ packages/plugin-core/src/commands/EnableTelemetry.ts
@@ -2,6 +2,7 @@ import { VSCodeEvents } from "@dendronhq/common-all";
 import { SegmentClient } from "@dendronhq/common-server";
 import { window } from "vscode";
 import { DENDRON_COMMANDS } from "../constants";
+import { isVSCodeTelemetryEnabled } from "../telemetry";
 import { AnalyticsUtils } from "../utils/analytics";
 import { BasicCommand } from "./base";
 
@@ -20,6 +21,8 @@ export class EnableTelemetryCommand extends BasicCommand<
     return {};
   }
   async execute() {
+    // If telemetry is disabled in vscode settings but users asks us to enable telemetry, force enable to override vscode
+    if (!isVSCodeTelemetryEnabled()) SegmentClient.forceEnable();
     SegmentClient.enable();
     AnalyticsUtils.track(VSCodeEvents.EnableTelemetry);
     window.showInformationMessage("telemetry enabled");
